abstract onPageNav, add to About
