+++ src/app/about/bio.mdx
@@ -1,0 +1,11 @@
+## Bio
+
+I grew up in North County San Diego, but spent the 15 years living in LA before moving to the city of subdued excitement (Bellingham, WA).
+
+Like many teenagers, I had very little idea what I wanted to do with my life. I've loved building and designing from a young age, and I spent my high school summers working in construction. I started a pre-architecture in college, but that was short-lived after the combination of basically flunking Calc II and discovering that graphic design was a thing. Thankfully, I had the self-awareness to know that my skills as a designer alone would likely not provide anything close to resembling a living wage I also studied business.
+
+In my early post-grad years I began teaching myself to code. While most of the engineers I work with would probably disagree, I’ve grown to be a capable front-end developer, contributing production-ready code to sites, products, and systems.
+
+I’ve spent the better part of the last decade at an enterprise software company, where I started as the first web designer and gradually (together with other many talented designers, under the leadership of [Juan Hidalgo](https://hidalgo.design)) built a [multi-disciplinary design org](https://liferay.design).
+
+At the beginning of 2024, working through burnout, I left my role as Head of Design to spend time with my family after the birth of our daughter.
