+++ src/components/recommends.tsx
@@ -4,6 +4,7 @@ import { RecommendsProps } from '@/lib/utilities/types';
 import Link from '../components/ui/link';
 import recommendsData from '@/lib/data/recommends.json';
 import { Card, CardHeader, CardTitle, CardContent } from './ui/card';
+import OnPageNav from './ui/on-page-nav';
 
 import {
   Tooltip,
@@ -12,8 +13,8 @@ import {
   TooltipTrigger,
 } from './ui/tooltip';
 import Image from 'next/image';
-import { ArrowDown, ArrowUp, ArrowRight } from 'lucide-react';
-import ProgressiveBlur from './ui/progressiveblur';
+// import { ArrowDown, ArrowUp, ArrowRight } from 'lucide-react';
+// import ProgressiveBlur from './ui/progressiveblur';
 
 const Recommends: React.FC<{ className?: string }> = ({ className }) => {
   const [recommends, setRecommends] = useState<RecommendsProps[]>([]);
@@ -40,10 +41,10 @@ const Recommends: React.FC<{ className?: string }> = ({ className }) => {
     window.scrollTo({ top, behavior: 'smooth' });
   };
 
-  // Create a state to store the direction of each arrow
-  const [arrowDirections, setArrowDirections] = useState<
-    Record<string, 'up' | 'down' | 'none'>
-  >({});
+  // // Create a state to store the direction of each arrow
+  // const [arrowDirections, setArrowDirections] = useState<
+  //   Record<string, 'up' | 'down' | 'none'>
+  // >({});
 
   // Create categoryRefs
   const categoryRefs = Object.keys(groupedRecommends).reduce(
@@ -55,93 +56,64 @@ const Recommends: React.FC<{ className?: string }> = ({ className }) => {
   );
 
   // Update arrowDirections when the scroll position changes
-  useEffect(() => {
-    const handleScroll = () => {
-      const newArrowDirections = Object.entries(categoryRefs).reduce(
-        (directions, [category, ref]) => {
-          if (ref.current) {
-            const rect = ref.current.getBoundingClientRect();
-            const middleOfElement = rect.top + rect.height / 2;
-            const middleOfViewport = window.innerHeight / 2;
-            if (middleOfElement >= 0 && middleOfElement <= window.innerHeight) {
-              directions[category] = 'none';
-            } else if (middleOfElement > middleOfViewport) {
-              directions[category] = 'down';
-            } else {
-              directions[category] = 'up';
-            }
-          }
-          return directions;
-        },
-        {} as Record<string, 'up' | 'down' | 'none'>
-      );
-      setArrowDirections(newArrowDirections);
-    };
+  // useEffect(() => {
+  //   const handleScroll = () => {
+  //     const newArrowDirections = Object.entries(categoryRefs).reduce(
+  //       (directions, [category, ref]) => {
+  //         if (ref.current) {
+  //           const rect = ref.current.getBoundingClientRect();
+  //           const middleOfElement = rect.top + rect.height / 2;
+  //           const middleOfViewport = window.innerHeight / 2;
+  //           if (middleOfElement >= 0 && middleOfElement <= window.innerHeight) {
+  //             directions[category] = 'none';
+  //           } else if (middleOfElement > middleOfViewport) {
+  //             directions[category] = 'down';
+  //           } else {
+  //             directions[category] = 'up';
+  //           }
+  //         }
+  //         return directions;
+  //       },
+  //       {} as Record<string, 'up' | 'down' | 'none'>
+  //     );
+  //     setArrowDirections(newArrowDirections);
+  //   };
 
-    // Nice Horizontal Arrow --> disappears when you get to end
-    window.addEventListener('scroll', handleScroll);
-    return () => window.removeEventListener('scroll', handleScroll);
-  }, [categoryRefs]);
-  const [arrowOpacity, setArrowOpacity] = useState(1);
-  const [arrowRotate, setArrowRotate] = useState(0); 
+  //   // Nice Horizontal Arrow --> disappears when you get to end
+  //   window.addEventListener('scroll', handleScroll);
+  //   return () => window.removeEventListener('scroll', handleScroll);
+  // }, [categoryRefs]);
+  // const [arrowOpacity, setArrowOpacity] = useState(1);
+  // const [arrowRotate, setArrowRotate] = useState(0); 
 
-  useEffect(() => {
-    const parentDiv = document.getElementById('onPageNav');
-    if (!parentDiv) return;
+  // useEffect(() => {
+  //   const parentDiv = document.getElementById('onPageNav');
+  //   if (!parentDiv) return;
 
-    const handleScroll = () => {
-      // Calculate the maximum scroll position
-      const maxScrollLeft = parentDiv.scrollWidth - parentDiv.clientWidth;
-      // Calculate the opacity based on the current scroll position and the maximum scroll position
-      const opacity = parentDiv.scrollLeft < maxScrollLeft ? 1 - parentDiv.scrollLeft / maxScrollLeft : 1;
-      const rotate = parentDiv.scrollLeft < maxScrollLeft ? 0 : 180;
-      // Set the opacity state variable to the calculated opacity
-      setArrowOpacity(opacity);
-      setArrowRotate(rotate);
-    };
+  //   const handleScroll = () => {
+  //     // Calculate the maximum scroll position
+  //     const maxScrollLeft = parentDiv.scrollWidth - parentDiv.clientWidth;
+  //     // Calculate the opacity based on the current scroll position and the maximum scroll position
+  //     const opacity = parentDiv.scrollLeft < maxScrollLeft ? 1 - parentDiv.scrollLeft / maxScrollLeft : 1;
+  //     const rotate = parentDiv.scrollLeft < maxScrollLeft ? 0 : 180;
+  //     // Set the opacity state variable to the calculated opacity
+  //     setArrowOpacity(opacity);
+  //     setArrowRotate(rotate);
+  //   };
 
-    parentDiv.addEventListener('scroll', handleScroll);
-    return () => parentDiv.removeEventListener('scroll', handleScroll);
-  }, []);
+  //   parentDiv.addEventListener('scroll', handleScroll);
+  //   return () => parentDiv.removeEventListener('scroll', handleScroll);
+  // }, []);
 
-  const iconStyles = {
-    className: 'invisible group-hover:visible opacity-80 absolute -right-4',
-    stroke: 'currentColor',
-  };
+  // const iconStyles = {
+  //   className: 'invisible group-hover:visible opacity-80 absolute -right-4',
+  //   stroke: 'currentColor',
+  // };
 
   return (
     <div className={`relative flex flex-col mt-40 2xl:mt-0 main-content ${className} `}>
-      <div
-        className={
-          '-ml-3 pl-4 sm:-ml-6 sm:px-9 2xl:px-4 py-2 rounded-xl sticky top-0 sm:top-10 z-30 w-[calc(100% + .75rem)] sm:col-span-4 flex overflow-x-scroll'
-        }
-        id="onPageNav"
-      >
-        {/* overflow-visible  */}
-        {/*  overflow-x-scroll overflow-y-clip */}
-        {Object.keys(groupedRecommends).map((category) => (
-          <button
-            className="group text-nowrap p-3 mr-8 flex relative"
-            onClick={() => scrollTo(category)}
-            key={category}
-          >
-            {category}
-            {arrowDirections[category] === 'up' ? (
-              <ArrowUp {...iconStyles} />
-            ) : arrowDirections[category] === 'down' ? (
-              <ArrowDown {...iconStyles} />
-            ) : null}
-          </button>
-        ))}
-        <div className="sm:hidden sticky rounded-md right-0 z-10 backdrop-blur-md my-auto p-3 flex">
-          <ArrowRight
-            size={24}
-            className="center-self-center opacity-80 transition-transform ease-in-out"
-            style={{ opacity: arrowOpacity, transform: `rotate(${arrowRotate}deg)` }}
-          />
-        </div>
-      </div>
-      <ProgressiveBlur className="block sticky -ml-4 sm:-ml-80 -mt-10 z-20 top-0 h-32 sm:h-64 w-dvw max-w-4xl 2xl:-ml-32 rotate-180" />
+      <OnPageNav categories={Object.keys(groupedRecommends)} scrollTo={scrollTo}/>
+      
       {Object.entries(groupedRecommends).map(([category, recommends]) => (
         <div
           className="my-16 sm:my-32 2xl:my-20"
@@ -247,3 +219,35 @@ export default Recommends;
 // TODO --> style current onPage items for mobile
 // TODO --> styles for the diff tags
 // TODO --> link tags? lots more work there
+
+// <div
+//         className={
+//           '-ml-3 pl-4 sm:-ml-6 sm:px-9 2xl:px-4 py-2 rounded-xl sticky top-0 sm:top-10 z-30 w-[calc(100% + .75rem)] sm:col-span-4 flex overflow-x-scroll'
+//         }
+//         id="onPageNav"
+//       >
+//         {/* overflow-visible  */}
+//         {/*  overflow-x-scroll overflow-y-clip */}
+//         {Object.keys(groupedRecommends).map((category) => (
+//           <button
+//             className="group text-nowrap p-3 mr-8 flex relative"
+//             onClick={() => scrollTo(category)}
+//             key={category}
+//           >
+//             {category}
+//             {arrowDirections[category] === 'up' ? (
+//               <ArrowUp {...iconStyles} />
+//             ) : arrowDirections[category] === 'down' ? (
+//               <ArrowDown {...iconStyles} />
+//             ) : null}
+//           </button>
+//         ))}
+//         <div className="sm:hidden sticky rounded-md right-0 z-10 backdrop-blur-md my-auto p-3 flex">
+//           <ArrowRight
+//             size={24}
+//             className="center-self-center opacity-80 transition-transform ease-in-out"
+//             style={{ opacity: arrowOpacity, transform: `rotate(${arrowRotate}deg)` }}
+//           />
+//         </div>
+//       </div>
+//       <ProgressiveBlur className="block sticky -ml-4 sm:-ml-80 -mt-10 z-20 top-0 h-32 sm:h-64 w-dvw max-w-4xl 2xl:-ml-32 rotate-180" />
